Fix charset to fit IANA standard
Per the character sets listed in the standard: http://www.iana.org/assignments/character-sets/character-sets.xhtml, 'utf8' is not the name of the character encoding. The correct encoding name is 'utf-8'. Correcting this brings it in conformance with web standards, allowing more interoperability. In particular .NET's Encoding.GetEncoding(string) function only returns an Encoding instance for 'utf-8' and null for 'utf8'.