Add new data field to DocumentLink to support documentLink/resolve properly
Without any context, it is not possible for a documentLink/resolve
request to resolve anything because all a DocumentLink has is a
non-null Range. To allow information to be transferred between a
textDocument/documentLink request and a documentLink/resolve request,
a new data field has been added for servers to populate it with
information necessary for the resolution of the link.

Signed-off-by: Remy Suen <remy.suen@gmail.com>