Add deprecated property to CompletionItem and SymbolInformation
It should be possible for completion items and symbols to be flagged
as being deprecated. Some editors and IDEs flag such deprecated items
by applying a visual affordance such as striking the text out. By
introducing this new deprecated property, these editors and IDEs will
be able to use its deprecation UI affordance when supporting a
language via the LSP.

Signed-off-by: Remy Suen <remy.suen@gmail.com>