Add kerboscript language server to server.md
Added the language server for kerboscript to the LSP implementations tables. 

This is a side project I've been working on for a programming language, for a mod of a game. The documentation for the kerboscript can be found [here](https://ksp-kos.github.io/KOS/). The main mod / programming language project can be found [here](https://github.com/KSP-KOS/KOS). 

Obviously this is a hyper niche language, and can understand not including to highlight useful languages.