update official graphql implementation entry
hey there!

just updating the entry for `graphql-language-service-server`, which has been merged into a larger monorepo, and is now an official project of the graphql foundation. 

it is used by the second most popular graphql language extension for vscode, as well as the most popular graphql completion for vim! we are almost feature complete, just need references/implementations. 

thanks for such an excellent protocol and implementation via `vscode-languageservice`!