Clarify SemanticTokenTypes.member is for functions
Many languages including C and C++ define "member" to include consider instance variables/fields/data members.
However VSCode maps `member` to the textmate scope `entity.name.function.member`, i.e. method/member function.
The intent seems to be that methods are `member` and fields are `property`.

It may be too late to change the names to be less ambiguous, but at least we can clarify in the spec.