Make versions consistently have type uinteger
`PublishDiagnosticParams` has its `version` field specified as a
`uinteger`, but other occurences of `version` are `integer`s.

This is inconsistent, and results in servers having to do
(theoretically) lossy conversions to `uinteger`s when creating
`PublishDiagnostics` notifications.

It seems to me that versions probably _should_ be `uintegers` (starting
at 0 and monotonically increasing), so this changes the other places to
use `uinteger`.

I suspect this will have no effect on implementations, but in theory
this could cause trouble for any clients who currently make use of
negative versions for something.