Grammar: i.e. not e.g.
The intended meaning is "UTF-8 code units (that is, bytes)", but e.g. is "for example" so is read as "UTF-8 code units (for example, bytes)". When I read this in the spec I was initially confused, does this flag mean _sometimes bytes and sometimes not_, because this is just an example? But then I realised it was just a grammatical error.